:recycle: use client useMarkNotificationsAs hook instead api call
